+++ api/lib/application/authentication/oidc/index.js
@@ -34,25 +34,6 @@ const register = async function (server) {
   server.route([
     ...adminRoutes,
     {
-      method: 'GET',
-      path: '/api/oidc/authorization-url',
-      config: {
-        auth: false,
-        validate: {
-          query: Joi.object({
-            identity_provider: Joi.string().required(),
-            audience: Joi.string().valid('app', 'admin').optional(),
-          }),
-        },
-        handler: oidcController.getAuthorizationUrl,
-        notes: [
-          "- Cette route permet de récupérer l'url d'autorisation du partenaire.\n" +
-            '- Elle retournera également les valeurs state et nonce.',
-        ],
-        tags: ['api', 'oidc', 'authentication'],
-      },
-    },
-    {
       method: 'POST',
       path: '/api/oidc/token',
       config: {
