+++ api/tests/identity-access-management/acceptance/application/oidc-provider.route.test.js
@@ -1,3 +1,5 @@
+import querystring from 'node:querystring';
+
 import { createServer, expect, generateValidRequestAuthorizationHeader } from '../../../test-helper.js';
 
 describe('Acceptance | Identity Access Management | Application | Route | oidc-provider', function () {
@@ -36,7 +38,7 @@ describe('Acceptance | Identity Access Management | Application | Route | oidc-p
   });
 
   describe('GET /api/oidc/redirect-logout-url', function () {
-    it('returns an object which contains the redirect logout url with HTTP status code 200', async function () {
+    it('returns an object which contains the redirect logout url with an HTTP status code 200', async function () {
       // given
       const options = {
         method: 'GET',
@@ -52,6 +54,36 @@ describe('Acceptance | Identity Access Management | Application | Route | oidc-p
       expect(response.result.redirectLogoutUrl).to.equal(
         'https://oidc.example.net/ea5ac20c-5076-4806-860a-b0aeb01645d4/oauth2/v2.0/logout?client_id=client',
       );
+    });
+  });
+
+  describe('GET /api/oidc/authorization-url', function () {
+    it('returns an object which contains the authentication url with an HTTP status code 200', async function () {
+      // given
+      const query = querystring.stringify({
+        identity_provider: 'OIDC_EXAMPLE_NET',
+        audience: 'app',
+      });
+
+      // when
+      const response = await server.inject({
+        method: 'GET',
+        url: `/api/oidc/authorization-url?${query}`,
+      });
+
+      // then
+      expect(response.statusCode).to.equal(200);
+
+      const redirectTargetUrl = new URL(response.result.redirectTarget);
+
+      expect(redirectTargetUrl.origin).to.equal('https://oidc.example.net');
+      expect(redirectTargetUrl.pathname).to.equal('/ea5ac20c-5076-4806-860a-b0aeb01645d4/oauth2/v2.0/authorize');
+      expect(redirectTargetUrl.searchParams.get('redirect_uri')).to.equal(
+        'https://app.dev.pix.org/connexion/oidc-example-net',
+      );
+      expect(redirectTargetUrl.searchParams.get('client_id')).to.equal('client');
+      expect(redirectTargetUrl.searchParams.get('response_type')).to.equal('code');
+      expect(redirectTargetUrl.searchParams.get('scope')).to.equal('openid profile');
     });
   });
 });
